column-lineage-api/pyproject.toml
[project]
name = "column-lineage-api"
version = "0.1.0"
description = "FastAPI backend for database view column lineage analysis"
authors = [
    {name = "Development Team", email = "dev@company.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.2",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.25",
    "snowflake-connector-python>=3.12.0",
    "snowflake-sqlalchemy>=1.6.0",
    "pandas>=1.5.3",
    "numpy>=1.26.4",
    "sqlglot>=25.0.0",
    "boto3>=1.40.34",
    "cognitojwt[async]>=1.4.1",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "hvac>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-benchmark>=4.0.0",
    "httpx>=0.27.0",
    "ruff>=0.9.3",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.hatch.build.targets.sdist]
include = [
    "/api",
    "/tests",
    "/README.md",
    "/CHANGELOG.md",
]

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["api"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
